Move to internal commands namespace
